ci: parallel suites, tested bundles, recoverable releases
CI and test-fast run under pytest-xdist (installed everywhere, used only
locally until now) with pip caching and CPU torch on Linux runners. The
628-test streamed oracle splits along its own axes so loadfile can
spread it. The five published apps/* bundles' test suites finally run in
CI and at publish, and apps joins the lint targets. konfai-mcp pins
konfai/konfai-apps to the exact scm version like every sibling instead
of floating over core internals; the 9-package upload gains
skip-existing so a partial publish is completable by re-run; the studio
wheel's built front is verified inside the artifact. Integration
subprocesses share one run_workflow helper with a timeout, the
duplicated API pipeline test is deleted, heavy imports skip instead of
erroring at collection, and the SimpleITK NumPy-2.5 warning flood is
filtered with its upstream note.
