+++ .github/workflows/konfai_apps_ci.yml
@@ -54,11 +54,15 @@ jobs:
         uses: actions/setup-python@ece7cb06caefa5fff74198d8649806c4678c61a1 # v6
         with:
           python-version: ${{ matrix.python-version }}
+          cache: "pip"
+          cache-dependency-path: pyproject.toml
 
       - name: Install core package
+        # Linux adds the CPU torch index: the default PyPI torch wheel bundles the multi-GB CUDA
+        # stack a CPU-only runner never uses (macOS/Windows PyPI wheels are already CPU-only).
         run: |
           python -m pip install -U pip
-          pip install -e ".[dev]"
+          pip install -e ".[dev]" ${{ runner.os == 'Linux' && '--extra-index-url https://download.pytorch.org/whl/cpu' || '' }}
 
       - name: Restore Hugging Face cache
         uses: actions/cache@caa296126883cff596d87d8935842f9db880ef25 # v5
@@ -73,3 +77,17 @@ jobs:
       - name: Run apps tests
         run: |
           python -m pytest -q konfai-apps/tests
+
+      - name: Install app bundles
+        # Editable, so each bundle's scm-derived `konfai==`/`konfai-apps==` pin resolves against
+        # the editable installs above (both sides derive the same version from this checkout).
+        shell: bash
+        run: |
+          for app in apps/*/; do pip install -e "./$app"; done
+
+      - name: Run app bundle tests
+        # The five PyPI-published bundles. impact_reg's integration test gates itself
+        # (KONFAI_IMPACTREG_REPO + fireants + CUDA) and skips here; its unit tests stub the runtime.
+        shell: bash
+        run: |
+          for app in apps/*/; do python -m pytest -q "${app}tests"; done
