+++ .github/workflows/konfai_ci.yml
@@ -45,15 +45,20 @@ jobs:
         uses: actions/setup-python@ece7cb06caefa5fff74198d8649806c4678c61a1 # v6
         with:
           python-version: ${{ matrix.python-version }}
+          cache: "pip"
+          cache-dependency-path: pyproject.toml
 
       - name: Install dependencies
+        # Linux adds the CPU torch index: the default PyPI torch wheel bundles the multi-GB CUDA
+        # stack a CPU-only runner never uses (macOS/Windows PyPI wheels are already CPU-only).
         run: |
           python -m pip install -U pip
-          pip install -e ".[dev]"
+          pip install -e ".[dev]" ${{ runner.os == 'Linux' && '--extra-index-url https://download.pytorch.org/whl/cpu' || '' }}
 
       - name: Run pytest
+        # Same xdist mode as the pixi `test` task; loadfile keeps per-file session fixtures on one worker.
         run: |
-          pytest -q tests
+          pytest -q -n auto --dist loadfile tests
 
   lint:
     runs-on: ubuntu-latest
@@ -68,6 +73,8 @@ jobs:
         uses: actions/setup-python@ece7cb06caefa5fff74198d8649806c4678c61a1 # v6
         with:
           python-version: "3.12"
+          cache: "pip"
+          cache-dependency-path: pyproject.toml
 
       - name: Install ruff
         run: pip install ruff==0.15.2
@@ -88,6 +95,8 @@ jobs:
         uses: actions/setup-python@ece7cb06caefa5fff74198d8649806c4678c61a1 # v6
         with:
           python-version: "3.12"
+          cache: "pip"
+          cache-dependency-path: pyproject.toml
 
       - name: Install ruff
         run: pip install ruff==0.15.2
@@ -109,13 +118,15 @@ jobs:
         uses: actions/setup-python@ece7cb06caefa5fff74198d8649806c4678c61a1 # v6
         with:
           python-version: "3.12"
+          cache: "pip"
+          cache-dependency-path: pyproject.toml
 
       - name: Install dependencies
         # The dev extra carries pytest + build; setuptools/setuptools-scm/wheel are the declared
         # build requirements the wheel test needs for a --no-isolation build.
         run: |
           python -m pip install -U pip
-          pip install -e ".[dev]" setuptools setuptools-scm wheel
+          pip install -e ".[dev]" setuptools setuptools-scm wheel --extra-index-url https://download.pytorch.org/whl/cpu
 
       - name: Build sdist and wheel
         run: python -m build --sdist --wheel
