+++ .github/workflows/konfai_mcp_ci.yml
@@ -39,11 +39,17 @@ jobs:
         uses: actions/setup-python@ece7cb06caefa5fff74198d8649806c4678c61a1 # v6
         with:
           python-version: ${{ matrix.python-version }}
+          cache: "pip"
+          cache-dependency-path: |
+            pyproject.toml
+            konfai-mcp/pyproject.toml
 
       - name: Install core package (with imaging backends)
+        # The CPU torch index: the default PyPI torch wheel bundles the multi-GB CUDA stack this
+        # CPU-only runner never uses.
         run: |
           python -m pip install -U pip
-          pip install -e ".[dev,imaging]"
+          pip install -e ".[dev,imaging]" --extra-index-url https://download.pytorch.org/whl/cpu
 
       - name: Install standalone apps package
         run: |
