+++ .github/workflows/konfai_studio_ci.yml
@@ -39,6 +39,10 @@ jobs:
         uses: actions/setup-python@ece7cb06caefa5fff74198d8649806c4678c61a1 # v6
         with:
           python-version: ${{ matrix.python-version }}
+          cache: "pip"
+          cache-dependency-path: |
+            pyproject.toml
+            studio/pyproject.toml
 
       - name: Set up Node
         uses: actions/setup-node@49933ea5288caeca8642d1e84afbd3f7d6820020 # v4
@@ -55,9 +59,11 @@ jobs:
           test -f ../konfai_studio/web/index.html
 
       - name: Install core package
+        # The CPU torch index: the default PyPI torch wheel bundles the multi-GB CUDA stack this
+        # CPU-only runner never uses.
         run: |
           python -m pip install -U pip
-          pip install -e ".[dev]"
+          pip install -e ".[dev]" --extra-index-url https://download.pytorch.org/whl/cpu
 
       - name: Install standalone apps package
         run: |
@@ -70,6 +76,18 @@ jobs:
       - name: Install standalone Studio package
         run: |
           pip install -e ./studio
+
+      - name: Build a throwaway studio wheel and verify it ships the front
+        # The wheel is the only artifact that carries web/ (git-ignored, swept in by package-data),
+        # so only inspecting a built wheel catches a packaging regression before release day.
+        run: |
+          python -m build --wheel studio
+          python -c "
+          import glob, sys, zipfile
+          names = zipfile.ZipFile(glob.glob('studio/dist/*.whl')[0]).namelist()
+          ok = 'konfai_studio/web/index.html' in names and any(n.startswith('konfai_studio/web/assets/') for n in names)
+          sys.exit(0 if ok else 'the studio wheel is missing konfai_studio/web/index.html or assets/')
+          "
 
       - name: Lint and format check
         run: |
