+++ konfai-mcp/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "konfai-mcp"
-dynamic = ["version"]
+dynamic = ["version", "dependencies"]
 description = "Standalone MCP server package for KonfAI experimentation workflows"
 readme = "README.md"
 license = "Apache-2.0"
@@ -14,21 +14,8 @@ authors = [
   { name = "Valentin Boussot", email = "boussot.v@gmail.com" }
 ]
 
-# konfai carries a bound because runner.py imports konfai.transformer at module scope: a 1.7.0
-# satisfies a bare name, and the server then dies on import -- all of it, not just the transform
-# tools. Spelled `>1.7.0` and not `>=1.8.0` so a source checkout satisfies it: setuptools_scm builds
-# the unreleased tree as 1.7.1.devN, which PEP 440 sorts BELOW 1.7.1 and above 1.7.0.
-# konfai-apps carries the same bound: its imports are lazy, so an older one does not break the server on
-# import, it breaks one app tool at the call with an ImportError naming a symbol instead of a version.
-dependencies = [
-  "konfai>1.7.0",
-  "konfai-apps>1.7.0",
-  # >= 2.10.2 for stateless streamable HTTP (stateless_http/json_response run
-  # kwargs and the FASTMCP_STATELESS_HTTP setting) per MCP spec rev 2025-03-26.
-  "fastmcp>=2.10.2",
-  "ruamel.yaml",
-  "numpy"
-]
+# Dependencies live in setup.py: konfai and konfai-apps are pinned to the exact scm-derived release
+# version there, like every other member of the lockstep family.
 
 [project.urls]
 Homepage = "https://github.com/fideus-labs/KonfAI"
