+++ pyproject.toml
@@ -176,12 +176,12 @@ scikit-image = "*"
 
 [tool.pixi.feature.dev.tasks]
 test = { cmd = "pytest -q -n auto --dist loadfile tests/", description = "Run the test suite" }
-test-fast = { cmd = 'pytest -q -m "not slow and not integration" tests/', description = "Run the test suite without slow oracle and integration tests" }
+test-fast = { cmd = 'pytest -q -n auto --dist loadfile -m "not slow and not integration" tests/', description = "Run the test suite without slow oracle and integration tests" }
 test-apps = { cmd = "pytest -q konfai-apps/tests", description = "Run the konfai-apps test suite" }
 test-cov = { cmd = "pytest -n auto --dist loadfile --cov=konfai --cov-report=term-missing tests/", description = "Run tests with coverage" }
-lint = { cmd = "ruff check konfai konfai-apps/konfai_apps konfai-mcp/konfai_mcp studio/konfai_studio tests", description = "Lint source code" }
-format = { cmd = "ruff format konfai konfai-apps/konfai_apps konfai-mcp/konfai_mcp studio/konfai_studio tests", description = "Format source code" }
-format-check = { cmd = "ruff format --check konfai konfai-apps/konfai_apps konfai-mcp/konfai_mcp studio/konfai_studio tests", description = "Check formatting without modifying files" }
+lint = { cmd = "ruff check konfai konfai-apps/konfai_apps konfai-mcp/konfai_mcp studio/konfai_studio apps tests", description = "Lint source code" }
+format = { cmd = "ruff format konfai konfai-apps/konfai_apps konfai-mcp/konfai_mcp studio/konfai_studio apps tests", description = "Format source code" }
+format-check = { cmd = "ruff format --check konfai konfai-apps/konfai_apps konfai-mcp/konfai_mcp studio/konfai_studio apps tests", description = "Check formatting without modifying files" }
 typecheck = { cmd = "python -m mypy konfai konfai-apps/konfai_apps konfai-mcp/konfai_mcp studio/konfai_studio --ignore-missing-imports --no-site-packages", description = "Type-check all packages" }
 build = { cmd = "python -m build", description = "Build sdist and wheel" }
 check = { depends-on = ["lint", "format-check", "test", "test-apps"], description = "Run all quality checks" }
@@ -243,6 +243,12 @@ addopts = "-q"
 markers = [
     "slow: marks tests as slow (deselect with -m 'not slow')",
     "integration: marks integration tests",
+]
+filterwarnings = [
+    # SimpleITK's GetImageFromArray/GetArrayViewFromImage assigns array.shape in place
+    # (SimpleITK/extra.py), which NumPy 2.5 deprecated: ~1000 copies per heavy file run bury every
+    # other warning. Drop this once the SimpleITK floor ships a reshape-based fix.
+    "ignore:Setting the shape on a NumPy array has been deprecated:DeprecationWarning:SimpleITK",
 ]
 
 # ---------------------------------------------------------------------------
